requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
fpdf==1.7.2
h11==0.14.0
httpcore==1.0.6
httpx==0.25.2
idna==3.10
python-dotenv==1.0.0
python-telegram-bot==20.7
sniffio==1.3.1
flask==2.0.1
werkzeug==2.0.1
gunicorn==20.1.0